[SNEAKER-176] Add build option to skip running unit test suite during builds work (#93)
This patch renames the previous Makefile build option fast to run-test, which adds greater flexibility and comes with the ability to skip building and running unit test suite all together. Now, if the option is unspecified, the unit test suite runs 3 times. If the value passed is 1, it will run exactly once, and finally, if value passed is 0, the unit test suite will be skipped altogether.